backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
